Personalizedmsg/requirements.txt
google-generativeai
pandas
scikit-learn
joblib
requests